Refactor LCP error handling: Introduce ErrorDetails component
* Add ErrorDetails and ErrorDetailsTooltip components for improved error display in LCP status.
* Update status.module.scss to reflect changes in error handling structure.
