+++ projects/plugins/boost/app/assets/src/js/features/lcp/status/error-details.module.scss
@@ -1,0 +1,57 @@
+@import '$css/main/variables';
+
+$grey: #8e8e8e;
+$black: #101517;
+$white: #fff;
+
+.error-tooltip-wrapper {
+	position: relative;
+	display: inline-block;
+
+	.jb-error-tooltip {
+		background-color: $black;
+		color: $white;
+		padding: 16px 24px;
+		border-radius: 4px;
+		font-size: 14px;
+		width: max-content;
+		max-width: 30rem;
+		position: absolute;
+		top: 100%;
+		left: 50%;
+		transform: translateX( -50% );
+		margin-top: 8px;
+		box-shadow: 0 1px 2px 0 rgba( 0, 0, 0, 0.05 );
+		z-index: 1000;
+
+		hr {
+			border-top: 1px solid $grey;
+			border-bottom: none;
+			border-left: none;
+			border-right: none;
+			margin: 12px 0;
+		}
+
+		&__header {
+			font-size: 1rem;
+			font-weight: 600;
+		}
+
+		&__row {
+			list-style: disc;
+			margin: 8px 0 8px 12px;
+		}
+
+		&__pointer {
+			width: 0;
+			height: 0;
+			border-left: 8px solid transparent;
+			border-right: 8px solid transparent;
+			border-bottom: 8px solid $black;
+			position: absolute;
+			top: -8px;
+			left: 50%;
+			transform: translateX( -50% );
+		}
+	}
+}
