+++ projects/plugins/boost/app/assets/src/js/features/lcp/status/status.module.scss
@@ -70,8 +70,3 @@
 		}
 	}
 }
-
-.error-tooltip-wrapper {
-	position: relative;
-	display: inline-block;
-}
