+++ projects/plugins/boost/app/assets/src/js/features/lcp/status/status.tsx
@@ -1,87 +1,10 @@
-import { __, _n, sprintf } from '@wordpress/i18n';
-import clsx from 'clsx';
-import InfoIcon from '$svg/info';
-import { createInterpolateElement, useState } from '@wordpress/element';
-import styles from './status.module.scss';
-import { useLcpState } from '../lib/stores/lcp-state';
 import TimeAgo from '$features/critical-css/time-ago/time-ago';
-import { Button } from '@automattic/jetpack-components';
-import { ErrorTooltip } from './error-tooltip';
-
-const ErrorDetails = () => {
-	const [ query ] = useLcpState();
-	const lcpState = query?.data;
-
-	if ( lcpState?.status !== 'analyzed' ) {
-		return null;
-	}
-
-	const pages = lcpState?.pages;
-	if ( ! pages || pages.length === 0 ) {
-		return null;
-	}
-
-	const errors = pages.filter( page => ( page?.errors?.length || 0 ) > 0 );
-	if ( errors.length === 0 ) {
-		return null;
-	}
-
-	const errorMessages = errors.flatMap( p => p.errors );
-
-	return createInterpolateElement(
-		sprintf(
-			// translators: %d is a number of pages which failed to be optimized
-			_n(
-				'%d page could not be optimized. <errorDetails />',
-				'%d pages could not be optimized. <errorDetails />',
-				errorMessages.length,
-				'jetpack-boost'
-			),
-			errorMessages.length
-		),
-		{
-			errorDetails: <ErrorDetailsTooltip />,
-		}
-	);
-};
-
-const ErrorDetailsTooltip = () => {
-	const [ isVisible, setIsVisible ] = useState( false );
-	const [ query ] = useLcpState();
-	const lcpState = query?.data;
-
-	if ( lcpState?.status !== 'analyzed' ) {
-		return null;
-	}
-
-	const pages = lcpState?.pages;
-	if ( ! pages || pages.length === 0 ) {
-		return null;
-	}
-
-	const errors = pages.filter( page => ( page?.errors?.length || 0 ) > 0 );
-	if ( errors.length === 0 ) {
-		return null;
-	}
-
-	const errorMessages = errors.flatMap( p => ( p.errors || [] ).map( e => e.message ) );
-
-	return (
-		<div className={ styles[ 'error-tooltip-wrapper' ] }>
-			<Button
-				className={ styles[ 'optimize-button' ] }
-				variant="link"
-				size="small"
-				weight="regular"
-				onClick={ () => setIsVisible( ! isVisible ) }
-				aria-expanded={ isVisible }
-			>
-				{ __( 'View details', 'jetpack-boost' ) }
-			</Button>
-			{ isVisible && <ErrorTooltip errors={ errorMessages } /> }
-		</div>
-	);
-};
+import InfoIcon from '$svg/info';
+import { __ } from '@wordpress/i18n';
+import clsx from 'clsx';
+import { useLcpState } from '../lib/stores/lcp-state';
+import { ErrorDetails } from './error-details';
+import styles from './status.module.scss';
 
 const Status: React.FC = () => {
 	const [ query ] = useLcpState();
